# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Transfer Sederhana",
        "description": "Dokumentasi API Transfer Sederhana, dibuat oleh Aditya Nanda Utama untuk keperluan seleksi BackEnd Programmer di bosCOD",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "aditya.nanda0030@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rekening-admin": {
            "get": {
                "tags": [
                    "RekeningAdmin"
                ],
                "summary": "All RekeningAdmin",
                "description": "All RekeningAdmin",
                "operationId": "allRekeningAdmin",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "RekeningAdmin"
                ],
                "summary": "Edit RekeningAdmin by id",
                "description": "Edit RekeningAdmin by id",
                "operationId": "editRekeningAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id RekeningAdmin",
                        "example": "1"
                    },
                    {
                        "name": "nama_bank",
                        "in": "query",
                        "example": "BNI"
                    },
                    {
                        "name": "atas_nama",
                        "in": "query",
                        "example": "Bos COD"
                    },
                    {
                        "name": "no_rekening",
                        "in": "query",
                        "example": "98301920232"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "RekeningAdmin"
                ],
                "summary": "Create RekeningAdmin",
                "description": "Create RekeningAdmin",
                "operationId": "createRekeningAdmin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nama_bank": {
                                        "type": "string",
                                        "example": "BNI"
                                    },
                                    "atas_nama": {
                                        "type": "string",
                                        "example": "Bos COD"
                                    },
                                    "no_rekening": {
                                        "type": "string",
                                        "example": "98301920232"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "RekeningAdmin"
                ],
                "summary": "Delete RekeningAdmin",
                "description": "Delete RekeningAdmin by id RekeningAdmin",
                "operationId": "deleteRekeningAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id RekeningAdmin",
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transfer": {
            "get": {
                "tags": [
                    "Transfer"
                ],
                "summary": "All Transfer",
                "description": "All Transfer",
                "operationId": "allTransfer",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transfer"
                ],
                "summary": "Create Transfer",
                "description": "Create Transfer",
                "operationId": "createTransfer",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nilai_transfer": {
                                        "type": "string",
                                        "example": "50000"
                                    },
                                    "bank_tujuan": {
                                        "type": "string",
                                        "example": "BCA"
                                    },
                                    "rekening_tujuan": {
                                        "type": "string",
                                        "example": "98301920232"
                                    },
                                    "atasnama_tujuan": {
                                        "type": "string",
                                        "example": "Fulan"
                                    },
                                    "bank_pengirim": {
                                        "type": "string",
                                        "example": "BNI"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}